Re-organize the run command files slightly and add solhint plugin as a default
